Restructure tasks packages
