+++ spark_movies_etl/tasks/abstract/transformation.py
@@ -1,0 +1,13 @@
+from abc import ABC, abstractmethod
+
+from pyspark.sql import DataFrame
+
+
+class AbstractTransformation(ABC):
+    """
+    Base class to define a DataFrame transformation.
+    """
+
+    @abstractmethod
+    def transform(self, df: DataFrame) -> DataFrame:
+        raise NotImplementedError
