+++ tests/spark_movies_etl/integration/fixtures/test_config.yaml
@@ -4,8 +4,8 @@ dynaconf_merge:
 
 global:
   task_argument_class_mapping:
-    standardize: spark_movies_etl.tasks.task_standardize_data.StandardizeDataTask
-    curate: spark_movies_etl.tasks.task_curate_data.CurateDataTask
+    standardize: spark_movies_etl.tasks.standardize_data.task.StandardizeDataTask
+    curate: spark_movies_etl.tasks.curate_data.task.CurateDataTask
   movie_languages_filter: ["EN"]
   data_lake:
     raw:
